+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Ability to download purchased vouchers
-- Portokasse journal lists top up history
+- Portokasse journal lists top up and order history
 
 ## [0.5.1] - 2021-08-11